Refactor Dockerfile to allow running DockerOperators without root
The problem is that the worker container needs to access the host's
/var/run/docker.sock. We can map it by creating a Docker volume, but then we
need to solve the access permissions. What we were doing is running the worker
container as root (i.e. a privileged container). This works, but then we have to
maintain two Dockerfiles, two containers, while making sure both contain the
same contents. This is suboptimal.

What I've done here is, when running the container, it checks if there exists a
`/var/run/docker.sock`. If so, it detects which group owns it and creates a new
group with the same GID inside the container, adding the airflow user to it. We
need to keep the same GID, otherwise the permissions won't work. After this, we
drop back to the unprivileged "airflow" user and run the rest of the code.

This feels a bit hackish, but is better than maintaining another container.