[#488] Create DAG that dumps our API DB to S3 monthly
The DAG whitelists which tables to dump, so we don't dump private data by
mistake. It uses the `pg_dump` software, which means the worker needs to have
`pg_dump` installed and compatible with the version of the Postgres instance
it's dumping (i.e. equal or superior version).

Fixes opentrials/opentrials#488