[#734] Add merge_identifiers_and_reindex as a separate DAG
This DAG isn't scheduled, so it only runs when other DAGs trigger it.
Potentially, we'll run it more than necessary. Consider the following case:

1. nct DAG finishes and triggers merge_identifiers_and_reindex
2. Before merge_identifiers_and_reindex finishes, euctr and ictrp finishes and
both trigger it

It'll finish the run triggered by nct, then will start the run triggered from
euctr, and then ictrp. Ideally, it would ignore all pending runs, and only run
the latest one. Unfortunately, the operator to do it is only available in
Airflow 1.8.0, so we'll have to upgrade.

opentrials/opentrials#734