[#734] Upgrade Airflow to 35e43f5
I would like to use a released version, but we can't because of a bug with the
HttpSensor that wasn't fixed in any released version
yet (https://issues.apache.org/jira/browse/AIRFLOW-1030).

We need to upgrade Airflow in the first place to be able to use the
`LatestOnlyOperator`.

opentrials/opentrials#734