Fix bug where we ran Airflow using root's $HOME
This causes problems because `pip` tries to check for a cache on the user's
`$HOME/.local`. As it's checking on root's $HOME, this fails because the
airflow user has no permissions to read that folder.

This is caused because we're calling airflow using `sudo -E`, which pass all of
our current user's environment variables. This is needed because we have to
pass forward things like S3 keys received via the environment. However, we also
end up passing the $HOME forward.

The fix is to add the `-H` flag to our `sudo` call, which changes the $HOME to
the target user's home.