[#812] Add `DockerCLIOperator` that runs containers using Docker CLI
Airflow's DockerOperator has an issue with long-running tasks, where it doesn't
detect their exit status, keeping the tasks stuck in a "running" state even
after they're finished (see
https://issues.apache.org/jira/browse/AIRFLOW-1131). This commit implements the
DockerCLIOperator that instead of using the Docker API, as the DockerOperator
does, uses the Docker CLI executable.

I have tested locally and it seems to work around the issues we've been having.
However, we'll only be sure after testing in production. This commit also
changes the EUCTR processor task to use the new DockerCLIOperator, so we can
try it. If it does work, we'll change the helpers to use DockerCLIOperator
instead of Airflow's DockerOperator.

opentrials/opentrials#812