[#812] Fix bug when passing env to DockerCLIOperator and log STDERR
The environment passed to `subprocess.Popen()` needs to:

* Have all values as strings;
* All strings must be byte strings.

We solved this by removing `None` values, converting everything else to string,
and encoding as UTF-8.

The other issue was that we were running Docker passing the env variables like
`--env "FOO=$FOO"`, expecting that the `$FOO` variable would take the contents
of the `FOO` env variable. It wasn't the case. To do so, we had to instead of
running the command directly, running it like `/bin/bash -c "command"`, this
way bash would parse and change the env variables.

The last issue was that we weren't logging the STDERR output, which was fixed
by simply piping it to STDOUT.

opentrials/opentrials#812