[#812] Use DockerCLIOperator instead of DockerOperator
The DockerOperator has a bug where it fails to notice when a long-running
container is stopped, keeping the Airflow tasks in "running" state (see
https://issues.apache.org/jira/browse/AIRFLOW-1131). To workaround this, I
wrote the DockerCLIOperator which, instead of using the `docker` Python
bindings, runs Docker directly from a `subprocess`. After a couple weeks
testing with EUCTR, this has been working fine, so now we're using this
operator everywhere.

opentrials/opentrials#812